+++ Danki2/Assets/Prefabs/UI Elements/AbilityTree/AbilityListingPanel.prefab
@@ -242,9 +242,9 @@ GameObject:
   - component: {fileID: 7179800382927097761}
   - component: {fileID: 1407149023251407613}
   - component: {fileID: 3946085311465520546}
-  - component: {fileID: 659691438823435890}
+  - component: {fileID: 6966794857247870339}
   m_Layer: 5
-  m_Name: AbilityPanel
+  m_Name: AbilityListingPanel
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -310,7 +310,7 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
---- !u!114 &659691438823435890
+--- !u!114 &6966794857247870339
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -319,7 +319,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 7859317892095288869}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e1d7770ed149a6842baee506f015b8fb, type: 3}
+  m_Script: {fileID: 11500000, guid: fa47c15d30f45df479cc702bf61046b9, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   iconPanelImage: {fileID: 4684333385211926195}
