+++ Danki2/Assets/Scripts/UI/Menus/AbilityTreeEditor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3e233c0b0fcfab04da80a92d4ffc5bf3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
