+++ Danki2/Assets/Scripts/UI/Menus/AbilityTreeEditor/AbilityPanel.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 9f70f797f9c5fe74c8cea08a57767f6b
+guid: fa47c15d30f45df479cc702bf61046b9
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
