+++ Danki2/Assets/Scripts/UI/Menus/AbilityTreeEditor/AbilityTreeDisplay.cs
@@ -1,22 +1,10 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class AbilityTreeMenuController : MonoBehaviour
+internal class AbilityTreeDisplay : MonoBehaviour
 {
-    private Player player;
-
-    private float numTreeVerticalSections;
-    private Dictionary<Node, float> sectionIndices = new Dictionary<Node, float>();
-
-    [SerializeField]
-    private GameObject abilityTreeListContent = null;
-
-    [SerializeField]
-    private AbilityPanel abilityListingPanelPrefab = null;
-
     [SerializeField]
     private Image rootNodeOrb = null;
 
@@ -28,37 +16,19 @@ public class AbilityTreeMenuController : MonoBehaviour
 
     [SerializeField]
     private TreeAbility treeAbilityPrefab = null;
-    private int maxTreeDepth;
 
-    void Start()
+    private Player player;
+    private float numTreeVerticalSections;
+    private int maxTreeDepth;
+    private Dictionary<Node, float> sectionIndices = new Dictionary<Node, float>();
+
+    public void Start()
     {
         player = RoomManager.Instance.Player;
-
-        PopulateAbilityList();
         PopulateTree();
     }
 
-    private void PopulateAbilityList()
-    {
-        //destroy all current panels
-        foreach (Transform child in abilityTreeListContent.transform)
-        {
-            Destroy(child.gameObject);
-        }
-
-        foreach (KeyValuePair<AbilityReference, int> item in player.AbilityInventory)
-        {
-            if (item.Value > 0)
-            {
-                AbilityPanel abilityPanel = Instantiate(abilityListingPanelPrefab, Vector3.zero, Quaternion.identity);
-                abilityPanel.transform.SetParent(abilityTreeListContent.transform, false);
-
-                abilityPanel.Initialise(item.Key, item.Value);
-            }
-        }
-    }
-
-    private void PopulateTree()
+    public void PopulateTree()
     {
         numTreeVerticalSections = 0;
         sectionIndices.Clear();
