+++ Danki2/Assets/Scripts/UI/Menus/AbilityTreeEditor/AbilityTreeDisplay.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e1d7770ed149a6842baee506f015b8fb
+guid: d7a30f5a132c8aa4789af6ca4793c5b2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
